+++ src/Component.js
@@ -3,299 +3,8 @@ import Model from './Model.js';
 import getResult from './utils/getResult.js';
 import deepFlat  from './utils/deepFlat.js';
 import parseHTML from './utils/parseHTML.js';
-import syncNode from './utils/syncNode.js';
 import getAttributesHTML from './utils/getAttributesHTML.js';
-import getAttributesDiff from './utils/getAttributesDiff.js';
-
-const SYNC_PROPS = ['value', 'checked', 'selected'];
-
-/**
- * Wrapper class for HTML strings marked as safe.
- * @param {string} value The HTML string to be marked as safe.
- * @property {string} value The HTML string.
- * @private
- */
-class SafeHTML {
-    constructor(value) {
-        this.value = value;
-    }
-
-    toString() {
-        return this.value;
-    }
-}
-
-/**
- * Wrapper class for partial templates that preserves structure.
- * @param {Array} items The items in the partial.
- * @private
- */
-class Partial {
-    constructor(items) {
-        this.items = items;
-    }
-}
-
-/**
- * Wrapper for interpolation results with markers.
- * @param {number} interpolationUid The interpolation UID for markers.
- * @param {any} result The interpolation result.
- * @private
- */
-class InterpolationWrapper {
-    constructor(interpolationUid, result) {
-        this.interpolationUid = interpolationUid;
-        this.result = result;
-    }
-}
-
-/**
- * Manager for delegated events.
- * @private
- */
-class EventsManager {
-    constructor() {
-        this.listeners = [];
-        this.types = new Set();
-        this.previousSize = 0;
-    }
-
-    /**
-     * Add a listener to the events manager.
-     * @param {Function} listener The listener to add.
-     * @param {string} type The type of event.
-     * @return {number} The index of the listener.
-     */
-    addListener(listener, type) {
-        this.types.add(type);
-        this.listeners.push(listener);
-        return this.listeners.length - 1;
-    }
-
-    /**
-     * Reset the events manager.
-     */
-    reset() {
-        this.listeners = [];
-        this.previousSize = this.types.size;
-    }
-
-    /**
-     * Check if there are pending types.
-     * @return {boolean} True if there are pending types, false otherwise.
-     */
-    hasPendingTypes() {
-        return this.types.size > this.previousSize;
-    }
-}
-
-/**
- * Manager for component position tracking and recycling.
- * @private
- */
-class PathManager {
-    constructor() {}
-
-    /**
-     * Reset before render.
-     */
-    reset() {
-        this.paused = false;
-        this.previous = this.tracked || new Map();
-        this.tracked = new Map();
-        this.positionStack = [0];
-    }
-
-    /**
-     * Push position to stack.
-     */
-    push() {
-        this.positionStack.push(0);
-    }
-
-    /**
-     * Pop position from stack.
-     */
-    pop() {
-        this.positionStack.pop();
-    }
-
-    /**
-     * Increment position.
-     */
-    increment() {
-        this.positionStack[this.positionStack.length - 1]++;
-    }
-
-    /**
-     * Pause tracking.
-     */
-    pause() {
-        this.paused = true;
-    }
-
-    /**
-     * Resume tracking.
-     */
-    resume() {
-        this.paused = false;
-    }
-
-    /**
-     * Get current path as array.
-     * @return {string} Current position path.
-     */
-    getPath() {
-        return this.positionStack.join('-');
-    }
-
-    /**
-     * Track component at current path.
-     * @param {Component} component The component to track.
-     * @return {Component} The component.
-     */
-    track(component) {
-        if (this.paused) return component;
-
-        this.tracked.set(
-            this.getPath(),
-            component
-        );
-
-        return component;
-    }
-
-    /**
-     * Find recyclable component by path and type.
-     * @param {Function} constructor The component constructor.
-     * @return {Component|null} The recyclable component or null.
-     */
-    findRecyclable(constructor) {
-        const prev = this.previous.get(this.getPath());
-        return prev && prev.constructor === constructor && !prev.key ? prev : null;
-    }
-}
-
-/**
- * Element reference for managing DOM element attributes.
- * @private
- */
-class Element {
-    /**
-     * Create an element reference.
-     * @param {Object} options The options object.
-     * @param {Function} options.getSelector Function that returns the CSS selector for the element.
-     * @param {Function} options.getAttributes Function that returns the attributes object for the element.
-     */
-    constructor(options) {
-        this.getSelector = options.getSelector;
-        this.getAttributes = options.getAttributes;
-        this.previousAttributes = {};
-    }
-
-    /**
-     * Attach the element reference to a DOM element.
-     * @param {Node} parent The parent node to search in.
-     */
-    hydrate(parent) {
-        this.ref = parent.querySelector(this.getSelector());
-    }
-
-    /**
-     * Update the element's attributes based on the difference with previous attributes.
-     */
-    update() {
-        // Attributes diff.
-        const attributes = this.getAttributes();
-        const { remove, add } = getAttributesDiff(attributes, this.previousAttributes);
-        // Store previous attributes.
-        this.previousAttributes = attributes;
-        // Remove attributes first so later `setAttribute` overrides if needed.
-        remove.forEach(attr => {
-            this.ref.removeAttribute(attr);
-            if (SYNC_PROPS.includes(attr) && attr in this.ref) {
-                // Reset property to default.
-                this.ref[attr] = attr === 'value' ? '' : false;
-            }
-        });
-        // Add / update attributes.
-        Object.keys(add).forEach(attr => {
-            const value = add[attr];
-            this.ref.setAttribute(attr, value);
-            if (SYNC_PROPS.includes(attr) && attr in this.ref) {
-                this.ref[attr] = attr === 'value' ? value : value !== false && value !== 'false';
-            }
-        });
-    }
-}
-
-/**
- * Interpolation reference for managing dynamic content between comment markers.
- * @private
- */
-class Interpolation {
-    /**
-     * Create an interpolation reference.
-     * @param {Object} options The options object.
-     * @param {Function} options.getUid Function that returns the unique identifier for the interpolation.
-     * @param {any} options.expression The expression to be evaluated for the interpolation.
-     */
-    constructor(options) {
-        this.getUid = options.getUid;
-        this.expression = options.expression;
-    }
-
-    /**
-     * Attach the interpolation reference to comment markers in the DOM.
-     * @param {Node} parent The parent node to search in.
-     */
-    hydrate(parent) {
-        const commentMap = new Map();
-        const walker = document.createTreeWalker(parent, NodeFilter.SHOW_COMMENT);
-
-        let node;
-        while ((node = walker.nextNode())) {
-            commentMap.set(node.textContent, node);
-        }
-
-        const uid = this.getUid();
-        const startMarker = Component.INTERPOLATION_START(uid);
-        const endMarker = Component.INTERPOLATION_END(uid);
-
-        this.ref = [
-            commentMap.get(startMarker), 
-            commentMap.get(endMarker)
-        ];
-    }
-
-    /**
-     * Update the interpolation content with a new fragment.
-     * @param {DocumentFragment} fragment The new content fragment to insert.
-     */
-    update(fragment) {
-        const [startComment, endComment] = this.ref;
-
-        const currentFirstElement = startComment.nextSibling;
-        const currentSingleChildElement = currentFirstElement.nextSibling === endComment;
-        const currentEmpty = currentFirstElement == endComment;
-        const fragmentChildren = fragment.children;
-
-        if (currentSingleChildElement && fragmentChildren.length === 1 && !currentFirstElement.getAttribute(Component.DATA_ATTRIBUTE_ELEMENT)) {
-            // There is a single child element that is not a component's root element. Sync node attributes and content.
-            syncNode(currentFirstElement, fragmentChildren[0]);
-        } else if (currentEmpty) {
-            // Interpolation is empty. Insert the fragment.
-            endComment.parentNode.insertBefore(fragment, endComment);
-        } else {
-            // Interpolation is not empty. Replace the interpolation content.
-            const range = document.createRange();
-            range.setStartAfter(startComment);
-            range.setEndBefore(endComment);
-            range.deleteContents();
-            range.insertNode(fragment);
-        }
-    }
-}
+import { SafeHTML, Partial, InterpolationWrapper, EventsManager, PathManager, Element, Interpolation } from './core/index.js';
 
 /**
  * Same as getResult, but pass context as argument to the expression.
@@ -621,11 +330,17 @@ const parseInterpolations = (template, expressions, interpolations) => {
             }
 
             const currentInterpolationUid = ++interpolationUid;
+
+            function getStart() {
+                return Component.INTERPOLATION_START(`${this.uid}-${currentInterpolationUid}`);
+            }
+            function getEnd() {
+                return Component.INTERPOLATION_END(`${this.uid}-${currentInterpolationUid}`);
+            }
             // Add interpolation reference to interpolations array.
             interpolations.push({
-                getUid : function() {
-                    return `${this.uid}-${currentInterpolationUid}`;
-                },
+                getStart,
+                getEnd,
                 expression : expressions[expressionIndex]
             });
             // Add new expression to expressions array.
@@ -1356,7 +1071,8 @@ class Component extends View {
                         getAttributes : element.getAttributes.bind(this)
                     })),
                     interpolations : interpolations.map(interpolation => new Interpolation({
-                        getUid : interpolation.getUid.bind(this),
+                        getStart : interpolation.getStart.bind(this),
+                        getEnd : interpolation.getEnd.bind(this),
                         expression : interpolation.expression
                     })),
                     parts,
