+++ src/core/index.js
@@ -1,0 +1,9 @@
+import SafeHTML from './SafeHTML.js';
+import Partial from './Partial.js';
+import InterpolationWrapper from './InterpolationWrapper.js';
+import EventsManager from './EventsManager.js';
+import PathManager from './PathManager.js';
+import Element from './Element.js';
+import Interpolation from './Interpolation.js';
+
+export { SafeHTML, Partial, InterpolationWrapper, EventsManager, PathManager, Element, Interpolation };
